Add topster edit functionality and PATCH API

+++ src/components/dialogs/dialog-delete-object.tsx
@@ -12,7 +12,9 @@ import { Topster } from '@/libs/interfaces/topster.interface'
 import { useDeleteJournal } from '@/hooks/use-journal'
 import { useRouter } from 'next/navigation'
 import { capitalizeFirstLetter } from '@/libs/utils/string'
-import { useDeleteTopster } from '@/hooks/use-topster'
+import { useDeleteTopster, useUpdateTopster } from '@/hooks/use-topster'
+import { TopsterUpdateRequest } from '@/libs/dto/topster.dto'
+import { JournalUpdateRequest } from '@/libs/dto/journal.dto'
 
 
 export interface DialogDeleteObjectProps {
@@ -20,22 +22,29 @@ export interface DialogDeleteObjectProps {
     onCloseAction: () => void
     object: Rating | Journal | Topster | null | undefined
     type?: 'rating' | 'journal' | 'topster'
+    action?: 'delete' | 'edit'
+    updateObject?: TopsterUpdateRequest | JournalUpdateRequest | null | undefined
 }
 
 export const DialogDeleteObject = ({
     open,
     onCloseAction,
     object,
-    type = 'rating'
+    type = 'rating',
+    action = 'delete',
+    updateObject = null
 }: DialogDeleteObjectProps) => {
     
     const appRouter = useRouter()
     const [password, setPassword] = useState<string>('')
+    
     const { mutateAsync: deleteRating, isPending: isRatingPending } = useDeleteRating()
     const { mutateAsync: deleteJournal, isPending: isJournalPending } = useDeleteJournal()
     const { mutateAsync: deleteTopster, isPending: isTopsterPending } = useDeleteTopster()
     
-    const isPending = useMemo(() => isRatingPending || isJournalPending || isTopsterPending, [isRatingPending, isJournalPending, isTopsterPending])
+    const { mutateAsync: updateTopster, isPending: isTopsterUpdatePending } = useUpdateTopster()
+    
+    const isPending = useMemo(() => isRatingPending || isJournalPending || isTopsterPending || isTopsterUpdatePending, [isRatingPending, isJournalPending, isTopsterPending, isTopsterUpdatePending])
     
     const handleDelete = useCallback(async () => {
         if (!object || isEmpty(password) || isPending) return
@@ -82,13 +91,52 @@ export const DialogDeleteObject = ({
         }
     }, [object, password, isPending, type, deleteRating, deleteJournal, deleteTopster, onCloseAction, appRouter])
     
+    const handleUpdate = useCallback(async () => {
+        if (!object || isEmpty(password) || isPending || !updateObject) return
+        try {
+            let response
+            switch (type) {
+                case 'journal':
+                /*response = await deleteJournal({
+                 id: object._id,
+                 journal: {
+                 password
+                 }
+                 })
+                 break*/
+                // TODO. Update journal logic
+                case 'topster':
+                    response = await updateTopster({
+                        id: object._id,
+                        topster: {
+                            ...updateObject,
+                            password
+                        }
+                    })
+                    break
+            }
+            if (response) {
+                setPassword('')
+                onCloseAction()
+                toast.success(`Updated ${type} successfully.`)
+                if (type === 'journal' || type === 'topster') {
+                    appRouter.back()
+                }
+            } else {
+                toast.error(`Failed to update ${type}. Please check your password and try again.`)
+            }
+        } catch {
+            toast.error(`Failed to update ${type}. Please check your password and try again.`)
+        }
+    }, [object, password, isPending, updateObject, type, updateTopster, onCloseAction, appRouter])
+    
     if (!object) return null
     return (
         <Dialog transition open={open} onClose={onCloseAction} className='relative z-50 transition duration-300 ease-out data-closed:opacity-0'>
             <div className='fixed inset-0 bg-black/50' aria-hidden='true'/>
             <div className='fixed inset-0 flex w-screen items-center justify-center p-4'>
                 <DialogPanel className='w-3/4 space-y-4 bg-[#33373B] text-white md:p-12 p-4 rounded-2xl flex flex-col'>
-                    <DialogTitle className='font-bold'>{`Delete ${capitalizeFirstLetter(type)}`}</DialogTitle>
+                    <DialogTitle className='font-bold'>{`${capitalizeFirstLetter(action)} ${capitalizeFirstLetter(type)}`}</DialogTitle>
                     <Input
                         value={password}
                         onChange={(e) => setPassword(e.target.value)}
@@ -97,10 +145,11 @@ export const DialogDeleteObject = ({
                         className='w-full py-1 border-white border'
                     />
                     <Button.Box
-                        text={`Delete ${capitalizeFirstLetter(type)}`}
+                        text={`${capitalizeFirstLetter(action)} ${capitalizeFirstLetter(type)}`}
                         disabled={isEmpty(password)}
                         onClick={() => {
-                            handleDelete().then(noop)
+                            if (action === 'delete') handleDelete().then(noop)
+                            handleUpdate().then(noop)
                         }}
                     />
                 </DialogPanel>
